+++ jimureport-example/src/main/resources/application-oracle.yml
@@ -1,0 +1,48 @@
+server:
+  port: 8085
+spring:
+  #配置静态资源
+  mvc:
+    static-path-pattern: /**
+  resource:
+    static-locations: classpath:/static/
+  #配置数据库
+  datasource:
+    url: jdbc:oracle:thin:@192.168.1.199:1521:helowin
+    username: jeecgbootbpm
+    password: jeecg196283
+    driver-class-name: oracle.jdbc.OracleDriver
+#Minidao配置
+minidao :
+  base-package: org.jeecg.modules.jmreport.*
+  #DB类型（mysql|postgresql|oracle|sqlserver|other）
+  db-type: sqlserver
+#JimuReport[上传配置]
+jeecg :
+  jmreport:
+    #自定义项目前缀
+    customPrePath:
+    #接口超时设置（毫秒）
+    connect-timeout: 300000
+  # local|minio|alioss
+  uploadType: local
+  # local
+  path :
+    #文件路径
+    upload: /opt/upload
+  # alioss
+  oss:
+    endpoint: ??
+    accessKey: ??
+    secretKey: ??
+    bucketName: jimureport
+  # minio
+  minio:
+    minio_url: http://minio.jeecg.com
+    minio_name: ??
+    minio_pass: ??
+    bucketName: ??
+#输出sql日志
+logging:
+  level:
+    org.jeecg.modules.jmreport : debug
