--- jimureport-samples/jimureport-demo/db/Dockerfile
@@ -1,9 +1,0 @@
-FROM mysql:8.0.19
-
-MAINTAINER jeecgos@163.com
-
-ENV TZ=Asia/Shanghai
-
-RUN ln -sf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
-
-COPY ./jimureport.mysql5.7.create.sql /docker-entrypoint-initdb.d
