--- jimureport-samples/jimureport-demo/pom.xml
@@ -1,182 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.0</version>
-        <relativePath/>
-    </parent>
-    <groupId>org.jeecg</groupId>
-    <artifactId>jimureport-demo</artifactId>
-    <version>1.3</version>
-    <name>jimureport-demo</name>
-
-    <url>http://www.jimureport.com</url>
-    <description>积木报表集成示例</description>
-
-    <repositories>
-        <repository>
-            <id>aliyun</id>
-            <name>aliyun Repository</name>
-            <url>http://maven.aliyun.com/nexus/content/groups/public</url>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-        </repository>
-        <repository>
-            <id>jeecg</id>
-            <name>jeecg Repository</name>
-            <url>http://maven.jeecg.org/nexus/content/repositories/jeecg</url>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-        </repository>
-    </repositories>
-
-    <properties>
-        <java.version>1.8</java.version>
-        <minio.version>8.0.3</minio.version>
-        <!-- DB驱动 -->
-        <postgresql.version>42.2.6</postgresql.version>
-        <ojdbc6.version>11.2.0.3</ojdbc6.version>
-        <sqljdbc4.version>4.0</sqljdbc4.version>
-        <mysql-connector-java.version>8.0.20</mysql-connector-java.version>
-        <sqlit.version>3.7.2</sqlit.version>
-        <hsqldb.version>2.2.8</hsqldb.version>
-        <h2.version>1.4.197</h2.version>
-        <derby.version>10.11.1.1</derby.version>
-    </properties>
-
-
-    <dependencies>
-        <!-- JimuReport -->
-        <dependency>
-            <groupId>org.jeecgframework.jimureport</groupId>
-            <artifactId>jimureport-spring-boot-starter</artifactId>
-            <version>1.3.7</version>
-        </dependency>
-
-
-        <!-- SpringBoot-->
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-freemarker</artifactId>
-        </dependency>
-
-        <!-- minio oss-->
-        <dependency>
-            <groupId>io.minio</groupId>
-            <artifactId>minio</artifactId>
-            <version>${minio.version}</version>
-            <optional>true</optional>
-        </dependency>
-
-        <!-- 数据库驱动 -->
-        <!--mysql-->
-        <dependency>
-            <groupId>mysql</groupId>
-            <artifactId>mysql-connector-java</artifactId>
-            <version>${mysql-connector-java.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--  sqlserver-->
-        <dependency>
-            <groupId>com.microsoft.sqlserver</groupId>
-            <artifactId>sqljdbc4</artifactId>
-            <version>${sqljdbc4.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!-- oracle驱动 -->
-        <dependency>
-            <groupId>com.oracle</groupId>
-            <artifactId>ojdbc6</artifactId>
-            <version>${ojdbc6.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!-- postgresql驱动 -->
-        <dependency>
-            <groupId>org.postgresql</groupId>
-            <artifactId>postgresql</artifactId>
-            <version>${postgresql.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!-- 达梦驱动 -->
-        <dependency>
-            <groupId>com.dameng</groupId>
-            <artifactId>DmJdbcDriver18</artifactId>
-            <version>1.0</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.dameng</groupId>
-            <artifactId>DmDialectForHibernate</artifactId>
-            <version>5.3</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!-- sqlite -->
-        <dependency>
-            <groupId>org.xerial</groupId>
-            <artifactId>sqlite-jdbc</artifactId>
-            <version>${sqlit.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--hsqldb-->
-        <dependency>
-            <groupId>org.hsqldb</groupId>
-            <artifactId>hsqldb</artifactId>
-            <version>${hsqldb.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--h2-->
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>${h2.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--derby-->
-        <dependency>
-            <groupId>org.apache.derby</groupId>
-            <artifactId>derbyclient</artifactId>
-            <version>${derby.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--db2-->
-        <dependency>
-            <groupId>com.ibm.db2</groupId>
-            <artifactId>jcc</artifactId>
-            <version>11.5.0.0</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--神通-->
-        <dependency>
-            <groupId>com.csicit.thirdparty</groupId>
-            <artifactId>oscar</artifactId>
-            <version>1.0.1</version>
-            <scope>runtime</scope>
-        </dependency>
-        <!--人大金仓-->
-        <dependency>
-            <groupId>kingbase</groupId>
-            <artifactId>kingbase8</artifactId>
-            <version>8</version>
-            <scope>runtime</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
