--- jimureport-samples/jimureport-demo/src/main/resources/application.yml
@@ -1,43 +1,0 @@
-server:
-  port: 8085
-spring:
-  #配置静态资源
-  mvc:
-    static-path-pattern: /**
-  resource:
-    static-locations: classpath:/static/
-  #配置数据库
-  datasource:
-    url: jdbc:mysql://${MYSQL-HOST:127.0.0.1}:${MYSQL-PORT:3306}/${MYSQL-DB:jimureport}?serverTimezone=UTC&useUnicode=true&characterEncoding=utf-8&allowPublicKeyRetrieval=true
-    username: root
-    password: root
-    driver-class-name: com.mysql.cj.jdbc.Driver
-#Minidao配置
-minidao :
-  base-package: org.jeecg.modules.jmreport.*
-  #DB类型（mysql|postgresql|oracle|sqlserver|other）
-  db-type: mysql
-#JimuReport[上传配置]
-jeecg :
-  # local|minio|alioss
-  uploadType: local
-  # local
-  path :
-    #文件路径
-    upload: /opt/upload
-  # alioss
-  oss:
-    endpoint: ??
-    accessKey: ??
-    secretKey: ??
-    bucketName: jimureport
-  # minio
-  minio:
-    minio_url: http://minio.jeecg.com
-    minio_name: ??
-    minio_pass: ??
-    bucketName: ??
-#输出sql日志
-logging:
-  level:
-    org.jeecg.modules.jmreport : info
